+++ src/main/java/racingcar/RacingGame.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.view.OutputView;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -7,7 +9,7 @@ public class RacingGame {
     private List<RacingCar> cars;
     private int finalTurn;
     private int turnCount;
-    private IOView ioView;
+    private OutputView outputView;
 
     public List<RacingCar> getCars() {
         return this.cars;
@@ -19,7 +21,7 @@ public class RacingGame {
 
     public RacingGame() {
         this.cars = new ArrayList<>();
-        this.ioView = new IOView();
+        this.outputView = new OutputView();
     }
     public RacingGame(List<String> names, int finalTurn) {
         this();
@@ -38,15 +40,15 @@ public class RacingGame {
             rc.accelerate(this.generateRandomNumber());
         }
         this.turnCount++;
-        ioView.printTurnResult(this.cars);
+        outputView.printTurnResult(this.cars);
     }
 
     public void playGame() {
-        ioView.printGameResultMessage();
+        outputView.printGameResultMessage();
         while(this.turnCount<this.finalTurn) {
             this.proceedTurn();
         }
-        ioView.printWinners(judgeWinners());
+        outputView.printWinners(judgeWinners());
     }
 
     public List<RacingCar> judgeWinners() {
